% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkg.R
\name{add_pkg}
\alias{add_pkg}
\title{Add package to named vector}
\usage{
add_pkg(pkgs, pkg)
}
\arguments{
\item{pkgs}{Processes \code{pkgs}, and \code{locals}, arguments of \code{\link{create_config}} and \code{\link{create_app}}.}

\item{pkg}{String. Name of package to add}
}
\value{
Package dependency string vector for \code{\link{standardize_pkgs}}.
}
\description{
Adds (named or not) package dependencies to a named vector of packages.
}
\author{
Jonathan Hill
}
