% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_pkgs}
\alias{add_pkgs}
\title{Add package to named vector}
\usage{
add_pkgs(pkgs, pkg)
}
\arguments{
\item{pkgs}{Processes \code{pkgs} argument of \code{\link{create_config}} and \code{\link{create_app}}.}

\item{pkg}{String. Name of package to add}
}
\value{
Package dependency string vector for \code{\link{standardize_pkgs}}.
}
\description{
Adds (named or not) package dependencies to a named vector of packages.
}
\examples{
pkgs <- c("shiny", rmarkdown = "1.8", "Rook")
add_pkgs(pkgs, c("rmarkdown", "dplyr"))

}
\author{
Jonathan Hill
}
\keyword{internal}
