% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code}
\alias{code}
\title{Pascal script to check registry for R}
\usage{
code(iss)
}
\arguments{
\item{iss}{Character vector which cummulatively becomes an Inno Setup Script (ISS).}
}
\value{
Chainable character vector, which can be used as the \code{text} argument of \code{\link{writeLines}} to generate an ISS.
}
\description{
Modern Delphi-like Pascal adds a lot of customization possibilities to the
installer. For examples, please visit \href{http://www.jrsoftware.org/ishelp/topic_scriptintro.htm}{Pascal Scripting Introduction}.
}
\details{
This script checks the registry for R, so that R will only be installed if necessary.
}
\examples{
\dontrun{
readLines(system.file('installation/code.iss', package = 'RInno'))
}

}
\seealso{
\code{\link{get_R}}, \code{\link{copy_installation}}, \code{\link{create_config}}, \code{\link{create_bat}}, \code{\link{directives}}, \code{\link{setup}}, \code{\link{languages}}, \code{\link{tasks}}, \code{\link{files}}, \code{\link{icons}}, \code{\link{run}}, and \code{\link{code}}.
}
\author{
Jonathan M. Hill
}
