% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_iss.R
\name{compile_iss}
\alias{compile_iss}
\title{Compile ISS}
\usage{
compile_iss()
}
\value{
Installer in \code{dir_out}.
}
\description{
After running \code{\link{create_app}} and editing the content of the
installer and app, call \code{compile_iss}.
}
\author{
Jonathan M. Hill
}
