% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_deployment.R
\name{copy_deployment}
\alias{copy_deployment}
\title{Default Deployment Files}
\usage{
copy_deployment(app_dir)
}
\arguments{
\item{app_dir}{Shiny app's directory, defaults to \code{getwd()}.}
}
\description{
This function moves files stored in \code{system.file('deployment', package = 'RInno')} to \code{app_dir}:
\itemize{
  \item Icons for installer and app, \emph{setup.ico} and \emph{default.ico}.
  \item Files that manage app start up, \emph{package_manager.R} and \emph{app.R}.
  \item First/last page of the installation wizard, \emph{infobefore.txt} and \emph{infoafter.txt}.
  \item Batch support files, \emph{wsf/run.wsf}, \emph{wsf/js/run.js}, \emph{wsf/js/json2.js}, and \emph{wsf/js/JSON.minify.js}.
}
}
\seealso{
\code{\link{create_app}}
}
\author{
Jonathan M. Hill
}
