% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cran_version}
\alias{cran_version}
\title{Check CRAN for package version}
\usage{
cran_version(pkg_name, cran_url = "http://cran.r-project.org/package=")
}
\arguments{
\item{pkg_name}{String. Package name as published on CRAN.}

\item{cran_url}{String. First part of the cannonical form of a package website on CRAN.}
}
\value{
The package's version as a \code{numeric_version}.
}
\description{
Check CRAN for package version
}
\examples{
cran_version("shiny")
}
\keyword{internal}
