% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\title{File Section of ISS}
\usage{
files(iss, app_dir, file_list = character())
}
\arguments{
\item{iss}{Character vector which cumulatively becomes an Inno Setup Script (ISS).}

\item{app_dir}{Shiny app's directory, defaults to \code{getwd()}.}

\item{file_list}{Character vector. Extra files to be installed with the app.}
}
\value{
Chainable character vector, which can be used as the \code{text} argument of \code{\link{writeLines}} to generate an ISS.
}
\description{
Files to be installed on user's computer. Everything in \code{app_dir} plus
\code{file_list}. For more information, visit \href{http://www.jrsoftware.org/ishelp/index.php?topic=filessection}{[Files] section}.
}
\author{
Jonathan M. Hill
}
