% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flexdashboard_check}
\alias{flexdashboard_check}
\title{Search for flexdashboard}
\usage{
flexdashboard_check(file_list)
}
\arguments{
\item{file_list}{Character vector. List of files within \code{app_dir}.}
}
\description{
This function locates a flexdashboard within a file list.
}
\author{
Hanjo Odendaal.
}
\keyword{internal}
