% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Chrome.R
\name{get_Chrome}
\alias{get_Chrome}
\title{Downloads Chrome}
\usage{
get_Chrome(app_dir)
}
\arguments{
\item{app_dir}{Shiny app's directory, defaults to \code{getwd()}.}
}
\value{
chrome_installer.exe in \code{app_dir}.
}
\description{
Downloads Chrome in \code{app_dir}. If Chrome has already been downloaded, \code{get_Chrome} will use that file. If the download fails it will stop.
}
\details{
If \code{\link{create_app}(include_Chrome = TRUE)}, then \code{get_Chrome}.
}
\seealso{
\code{\link{get_R}}, \code{\link{copy_installation}}, \code{\link{create_config}}, \code{\link{create_bat}}, \code{\link{directives}}, \code{\link{setup}}, \code{\link{languages}}, \code{\link{tasks}}, \code{\link{files}}, \code{\link{icons}}, \code{\link{run}}, and \code{\link{code}}.
}
\author{
Jonathan M. Hill
}
