% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_nodejs.R
\name{install_nodejs}
\alias{install_nodejs}
\title{Downloads and installs nodejs}
\usage{

  install_nodejs(page_with_download_url = "https://nodejs.org/en/download/",
  version = "LTS", ...)
}
\arguments{
\item{page_with_download_url}{nodejs download url.}

\item{version}{character. "current" or "lts". Defaults to "lts"}

\item{...}{extra parameters to pass to \code{\link[installr]{install.URL}}}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Suports Nodejs's "current" and "lts" versions - \href{https://nodejs.org/en/download/}{LTS} - \href{https://nodejs.org/en/download/current}{Current}
}
\details{
As an asynchronous event driven JavaScript runtime, Node is designed to build scalable network applications.

See \href{https://nodejs.org/en/about/}{About} for more information.
}
\examples{

\dontrun{
install_nodejs()
install_nodejs(version = "current")
}

}
\author{
Tal Galili, A. Jonathan R. Godfrey, and Jonathan M. Hill
}
