% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inno_doc.R
\name{ispp_doc}
\alias{ispp_doc}
\title{Inno Setup Preprossor Help}
\usage{
ispp_doc()
}
\value{
Local Inno Setup Preprocessor Documentation.
}
\description{
Inno Setup Preprossor Help
}
\author{
Jonathan M. Hill
}
