% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages_section.R
\name{languages_section}
\alias{languages_section}
\title{Languages Section of ISS}
\usage{
languages_section(iss, language = "english")
}
\arguments{
\item{iss}{Character vector which cumulatively becomes an Inno Setup Script (ISS).}

\item{language}{Character vector of lower case languages to include.}
}
\value{
Chainable character vector, which can be used as the \code{text} argument of \code{\link{writeLines}} to generate an ISS.
}
\description{
RInno currently supports 25 languages. Check the \code{languages} directory of Inno Setup for a complete list, and see \href{http://www.jrsoftware.org/ishelp/topic_languagessection.htm}{[Languages] section} for details.
}
\author{
Jonathan M. Hill
}
