% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativefy_app.R
\name{nativefy_app}
\alias{nativefy_app}
\title{Package app into electron with nativefier}
\usage{
nativefy_app(app_name, app_dir, nativefier_opts,
  app_icon = "default.ico")
}
\arguments{
\item{app_name}{The name of the app. It will be displayed throughout the installer's window titles, wizard pages, and dialog boxes. See \href{http://www.jrsoftware.org/ishelp/topic_setup_appname.htm}{[Setup]:AppName} for details. For continuous installations, \code{app_name} is used to check for an R package of the same name, and update it. The Continuous Installation vignette has more details.}

\item{app_dir}{Development app's directory, defaults to \code{getwd()}.}

\item{nativefier_opts}{Character vector. Extra options to pass to nativefier when \code{user_browser = "electron"}. Each string in the vector should be a valid nativefier command. For example, \code{c('--no-overwrite', '--conceal', '--show-menu-bar')}. For more information, \code{system("nativefier --help")}.}

\item{app_icon}{Filename of icon in \code{app_dir}, used for desktop and program menu shortcuts.}
}
\description{
Package app into electron with nativefier
}
