% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_R_version}
\alias{sanitize_R_version}
\title{Sanitize R's version}
\usage{
sanitize_R_version(R_version, clean = FALSE)
}
\arguments{
\item{R_version}{R version to use. Supports inequalities similar to \code{pkgs}. Defaults to: \code{paste0(">=", R.version$major, '.', R.version$minor)}.}

\item{clean}{Boolean. If TRUE, \code{><=} are removed. Defaults to FALSE.}
}
\description{
Used to validate R versions and strip off inequalities when necessary.
}
\keyword{internal}
