% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_pkgs.R
\name{standardize_pkgs}
\alias{standardize_pkgs}
\title{Standardize package dependencies}
\usage{
standardize_pkgs(pkgs)
}
\arguments{
\item{pkgs}{Processes \code{pkgs}, \code{remotes} and \code{locals}, arguments of \code{\link{create_config}} and \code{\link{create_app}}.}
}
\description{
Standardizes (named or not) character vectors of package dependencies. This creates a standard format for package dependency information stored in config.cfg
}
\author{
William Bradley
}
