% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_iss.R
\name{start_iss}
\alias{start_iss}
\title{Start ISS}
\usage{
start_iss(app_name)
}
\arguments{
\item{app_name}{The name of the app being installed. It will be displayed throughout the installer and uninstaller in window titles, wizard pages, and dialog boxes. See \href{http://www.jrsoftware.org/ishelp/topic_setup_appname.htm}{[Setup]:AppName} for details. For continuous deployments, \code{app_name} is used to check for an R package of the same name, and update it. The Deployment vignette has more details.}
}
\value{
\code{app_name} and set \code{options('Rinno.app_name' = app_name)}
}
\description{
Chain \code{\link{directives}} against this function to start building custom
installers.
}
\examples{
\dontrun{
start_iss('myapp') \%>\%
  directives(include_R = FALSE, R_version = '3.3.2')
}

}
\seealso{
\code{\link{directives}}.
}
\author{
Jonathan M. Hill
}
