% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tasks}
\alias{tasks}
\title{Tasks Section of ISS}
\usage{
tasks(iss, desktop_icon = TRUE)
}
\arguments{
\item{iss}{Character vector which cummulatively becomes an Inno Setup Script (ISS).}

\item{desktop_icon}{Logical. If TRUE, create a desktop shortcut.}
}
\value{
Chainable character vector, which can be used as the \code{text} argument of \code{\link{writeLines}} to generate an ISS.
}
\description{
Defines all of the user-customizable tasks during installation. These tasks appear as check boxes and radio buttons on the \emph{Select Additional Tasks} installer page. See \href{http://www.jrsoftware.org/ishelp/topic_taskssection.htm}{[Tasks] section} for details.
}
\seealso{
\code{\link{get_R}}, \code{\link{copy_installation}}, \code{\link{create_config}}, \code{\link{create_bat}}, \code{\link{create_pkgs}}, \code{\link{directives}}, \code{\link{setup}}, \code{\link{languages}}, \code{\link{tasks}}, \code{\link{files}}, \code{\link{icons}}, \code{\link{run}}, and \code{\link{code}}.
}
\author{
Jonathan M. Hill
}
