% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_app.R
\name{example_app}
\alias{example_app}
\title{Example app}
\usage{
example_app(wd)
}
\arguments{
\item{wd}{Path to working directory.}
}
\value{
Shiny app example.
}
\description{
Creates a basic app to test in \code{wd/app}.
}
\examples{
\dontrun{
example_app(getwd())
}

}
\author{
Jonathan M. Hill
}
