% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_app.R
\name{example_app}
\alias{example_app}
\title{Example app}
\usage{
example_app(app_dir = "app", wd = getwd())
}
\arguments{
\item{app_dir}{Shiny app's directory, defaults to "app".}

\item{wd}{Path to working directory. Defaults to \code{getwd()}.}
}
\value{
Shiny app example.
}
\description{
Creates a basic app to test in \code{wd/app_dir}.
}
\examples{
\dontrun{
example_app(getwd())
}

}
\author{
Jonathan M. Hill
}
