% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_section.R
\name{files_section}
\alias{files_section}
\title{Files Section of ISS}
\usage{
files_section(iss, app_name, app_dir, user_browser,
  file_list = character())
}
\arguments{
\item{iss}{Character vector which cumulatively becomes an Inno Setup Script (ISS).}

\item{app_name}{The name of the app. It will be displayed throughout the installer's window titles, wizard pages, and dialog boxes. See \href{http://www.jrsoftware.org/ishelp/topic_setup_appname.htm}{[Setup]:AppName} for details. For continuous installations, \code{app_name} is used to check for an R package of the same name, and update it. The Continuous Installation vignette has more details.}

\item{app_dir}{Development app's directory, defaults to \code{getwd()}.}

\item{user_browser}{Character for the default browser. Options include "chrome", "firefox", and "ie."}

\item{file_list}{Character vector. Extra files to be installed with the app.}
}
\value{
Chainable character vector, which can be used as the \code{text} argument of \code{\link{writeLines}} to generate an ISS.
}
\description{
Files to be installed on user's computer. Everything in \code{app_dir} plus
\code{file_list}. For more information, visit \href{http://www.jrsoftware.org/ishelp/index.php?topic=filessection}{[Files] section}.
}
\author{
Jonathan M. Hill
}
