% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Chrome.R
\name{get_Chrome}
\alias{get_Chrome}
\title{Downloads Chrome}
\usage{
get_Chrome(app_dir)
}
\arguments{
\item{app_dir}{Development app's directory, defaults to \code{getwd()}.}
}
\value{
chrome_installer.exe in \code{app_dir}.
}
\description{
Downloads Chrome in \code{app_dir}. If Chrome has already been downloaded, \code{get_Chrome} will use that file. If the download fails it will stop.
}
\details{
If \code{\link{create_app}(include_Chrome = TRUE)}, then \code{get_Chrome}.
}
\seealso{
\code{\link{get_R}}, \code{\link{copy_installation}}, \code{\link{create_config}}, \code{\link{create_bat}}, \code{\link{directives_section}}, \code{\link{setup_section}}, \code{\link{languages_section}}, \code{\link{tasks_section}}, \code{\link{files_section}}, \code{\link{icons_section}}, \code{\link{run_section}}, and \code{\link{code_section}}.
}
\author{
Jonathan M. Hill
}
