% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_R.R
\name{get_R}
\alias{get_R}
\title{Downloads R}
\usage{
get_R(app_dir = getwd(), R_version = paste0(">=", R.version$major, ".",
  R.version$minor))
}
\arguments{
\item{app_dir}{Development app's directory, defaults to \code{getwd()}.}

\item{R_version}{R version to use. Supports inequalities. Defaults to: \code{paste0(">=", R.version$major, '.', R.version$minor)}.}
}
\value{
\code{sprintf('R-\%s-win.exe', R_version)} in \code{app_dir}.
}
\description{
Downloads R in \code{app_dir}. If it has already been downloaded, \code{get_R} will use that file. If the download fails it will stop.
}
\details{
If \code{\link{create_app}(include_R = TRUE)}, then \code{get_R}.
}
\seealso{
\code{\link{get_R}}, \code{\link{copy_installation}}, \code{\link{create_config}}, \code{\link{create_bat}}, \code{\link{directives_section}}, \code{\link{setup_section}}, \code{\link{languages_section}}, \code{\link{tasks_section}}, \code{\link{files_section}}, \code{\link{icons_section}}, \code{\link{run_section}}, and \code{\link{code_section}}.
}
\author{
Jonathan M. Hill
}
