% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_inno.R
\name{install_inno}
\alias{install_inno}
\title{Downloads and installs Inno Setup}
\usage{
install_inno(quick_start_pack = FALSE, ...)
}
\arguments{
\item{quick_start_pack}{The Inno Setup QuickStart Pack includes Inno Setup and Inno Script Studio script editor. See \href{http://www.jrsoftware.org/is3rdparty.php}{Third-Party Files} page for more information.}

\item{...}{extra parameters to pass to \code{\link[installr]{install.URL}}}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Downloads and installs Inno Setup's \href{http://www.jrsoftware.org/isdl.php#stable}{stable release}
}
\details{
Inno Setup is a free installer for Windows programs. First introduced in 1997, it currently rivals many commercial installers in feature set and stability.

See \href{http://www.jrsoftware.org/isinfo.php#features}{Features} for more information.
}
\examples{

\dontrun{
install_inno()
install_inno(quick_start_pack = T)
}

}
\author{
Tal Galili and Jonathan M. Hill
}
