% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_pkgs}
\alias{standardize_pkgs}
\title{Standardize package dependencies}
\usage{
standardize_pkgs(pkgs)
}
\arguments{
\item{pkgs}{Processes \code{pkgs}, and \code{pkgs}, arguments of \code{\link{create_config}} and \code{\link{create_app}}.}
}
\value{
Package dependency list
}
\description{
Standardizes (named or not) character vectors of package dependencies and formats it for config.cfg.
}
\author{
William Bradley and Jonathan Hill
}
\keyword{internal}
